Not pulling same HEAD hash twice
Before, if current package.json is equal to the one at HEAD, it would be
attempted to be pulled again.
This was pretty harmless, just annoying.
But it revealed a small-window race condition, when on second pull, bundle could appear.
This lead to AssertError, because veendor tried to install an empty
diff.